Update of the parameters dictionnary
Homogeneous use of the parameters dictionnary in the code

Main function for Video object in Video.py is done

Fusionned Zoom and Translate function in Video.py because they both need to call and create an IMAGE object
and I wanted to be able to use both of them in the same video

Made change in explosion. Depending on what it will look like, I might split grain rotation and explosion into 2 animations
otherwise we might not be able to see artist image behing fractal.